+++ server-side/api/v1/views/users.py
@@ -3,12 +3,9 @@ Modules for users endpoints
 """
 from api.v1.views import app_views
 from flask import abort, jsonify, request, abort
-from dotenv import load_dotenv
-from os import getenv
+import os
+from uuid import uuid4
 from bcrypt import hashpw, gensalt, checkpw
-
-
-load_dotenv()
 
 
 @app_views.route('/users', methods=['GET'], strict_slashes=False)
@@ -55,9 +52,13 @@ def get_user_by_id(user_id):
                          "created_at": user.created_at, "updated_at": user.updated_at}
             return jsonify(user_data), 200
         with db:
+            auth_user = request.current_user
             user = db.find_user_by_id(user_id)
             if user is None:
                 abort(404)
+            if auth_user is not None:
+                if auth_user.user_id != user.user_id:
+                    return jsonify({"error": "Unauthorized"}), 401
             db.save()
             return jsonify(user.to_dict()), 200
     except Exception as e:
@@ -74,8 +75,53 @@ def post_user():
     """
     from models.db import DBStorage
     db = DBStorage()
-    user_data = request.json
+    user_data = {}
 
+    user_fullname = request.form.get('fullname')
+    user_verified = request.form.get('verified')
+    user_email = request.form.get('email')
+    user_password = request.form.get('password')
+    user_gender = request.form.get('gender')
+    user_phone1 = request.form.get('phone1')
+    user_phone2 = request.form.get('phone2')
+    user_about = request.form.get('about')
+    user_address = request.form.get('address')
+    user_town = request.form.get('town')
+    user_city = request.form.get('city')
+    user_state = request.form.get('state')
+    
+    if user_fullname is None:
+        return jsonify({"error": "fullname missing"}), 400
+    if user_email is None:
+        return jsonify({"error": "email missing"}), 400
+    if user_password is None:
+        return jsonify({"error": "password missing"}), 400
+    if user_gender is None:
+        return jsonify({"error": "gender missing"}), 400
+    if user_phone1 is None:
+        return jsonify({"error": "phone1 missing"}), 400
+    if user_about is None:
+        return jsonify({"error": "about missing"}), 400
+    if user_address is None:
+        return jsonify({"error": "address missing"}), 400
+    if user_town is None:
+        return jsonify({"error": "town missing"}), 400
+    if user_city is None:
+        return jsonify({"error": "city missing"}), 400
+    if user_state is None:
+        return jsonify({"error": "state missing"}), 400
+    user_data['fullname'] = user_fullname
+    user_data['email'] = user_email
+    user_data['verified'] = user_verified
+    user_data['password'] = user_password
+    user_data['gender'] = user_gender
+    user_data['phone1'] = user_phone1
+    user_data['phone2'] = user_phone2
+    user_data['about'] = user_about
+    user_data['address'] = user_address
+    user_data['town'] = user_town
+    user_data['city'] = user_city
+    user_data['state'] = user_state
     try:
         with db:
             hashed_password = hashpw(
@@ -106,14 +152,14 @@ def login_user():
         with db:
             user = db.find_user_by_email(user_email)
             if user is None:
-                return jsonify({"error": "No user found for that email"}), 404
+                return jsonify({"error": "Invalid login details"}), 401
             if not checkpw(user_password.encode('utf-8'), user.password.encode('utf-8')):
-                return jsonify({"error": "wrong password"}), 401
+                return jsonify({"error": "Invalid login details"}), 401
             from api.v1.auth.session_db_auth import SessionDBAuth
             session_auth = SessionDBAuth()
             session_id = session_auth.create_session(user.user_id)
             response = jsonify(user.to_dict())
-            cookie_name = getenv('SESSION_NAME')
+            cookie_name = os.getenv('SESSION_NAME')
             response.set_cookie(cookie_name, session_id)
             db.save()
             return response, 200
@@ -154,17 +200,21 @@ def put_user(user_id):
         if new_password is None:
             return jsonify({"error": "new_password missing"}), 400
         with db:
+            auth_user = request.current_user
             user = db.find_user_by_id(user_id)
             if user is None:
                 abort(404)
             if not checkpw(user_password.encode('utf-8'), user.password.encode('utf-8')):
                 return jsonify({"error": "wrong password"}), 401
+            if auth_user is not None:
+                if auth_user.user_id != user.user_id:
+                    return jsonify({"error": "Unauthorized"}), 401
             hashed_password = hashpw(new_password.encode('utf-8'), gensalt())
             user.password = hashed_password
             db.save()
             return jsonify(user.to_dict()), 201
     except Exception as e:
-        abort(404)
+        abort(401)
 
 
 @app_views.route('/users/<string:user_id>', methods=['DELETE'], strict_slashes=False)
@@ -176,15 +226,46 @@ def delete_user(user_id):
     db = DBStorage()
     try:
         with db:
+            auth_user = request.current_user
             user = db.find_user_by_id(user_id)
+            sess = db.find_session_by_id_by_user_id(user_id)
+            
             if user is None:
-                abort(404)
+                return jsonify({"error": "User not found"}), 404
+            if sess is None:
+                return jsonify({"error": "No session found for that user"}), 404
+            if auth_user is not None:
+                if auth_user.user_id != user.user_id:
+                    return jsonify({"error": "Unauthorized"}), 401
+
+            db.delete_session(sess)
             db.delete_user(user)
             db.save()
-            return jsonify({}), 200
+            return jsonify({}), 204  # 204 No Content for successful deletion
     except Exception as e:
-        abort(404)
+        return jsonify({"error": "Internal Server Error"}), 500
 
+
+def save_picture(form_picture):
+    """
+    - Save the picture
+    """
+    pics_id = str(uuid4())
+    _, f_ext = os.path.splitext(form_picture.filename)
+    
+    allowed_extensions = {'.png', '.jpg', '.jpeg'}
+    
+    if f_ext.lower() in allowed_extensions:
+        picture_fn = pics_id + f_ext
+        picture_path = os.path.join(os.path.dirname(__file__), '..', 'client-side', 'assets', 'profile_pics', picture_fn)
+        
+        # Ensure the directory exists before saving
+        os.makedirs(os.path.dirname(picture_path), exist_ok=True)
+        
+        form_picture.save(picture_path)
+        return picture_fn
+    else:
+        return None
 
 @app_views.route('/users/<string:user_id>', methods=['PUT'], strict_slashes=False)
 def update_user_by_user_id(user_id):
@@ -193,45 +274,49 @@ def update_user_by_user_id(user_id):
     """
     from models.db import DBStorage
     db = DBStorage()
-    user_data = request.json
-    if user_data is None:
-        return jsonify({"error": "Not a JSON"}), 400
-    if 'about' not in user_data:
-        return jsonify({"error": "about missing"}), 400
-    if 'address' not in user_data:
-        return jsonify({"error": "address missing"}), 400
-    if 'city' not in user_data:
-        return jsonify({"error": "city missing"}), 400
-    if 'town' not in user_data:
-        return jsonify({"error": "town missing"}), 400
-    if 'state' not in user_data:
-        return jsonify({"error": "state missing"}), 400
-    if 'gender' not in user_data:
-        return jsonify({"error": "missing gender"})
-    if 'fullname' not in user_data:
-        return jsonify({"error": "missing fullname"})
-    if 'phone1' not in user_data:
-        return jsonify({"error": "missing phone1"})
-    if 'email' not in user_data:
-        return jsonify({"error": "missing email"})
-    if 'verified' not in user_data:
-        return jsonify({"error": "missing verified"})
+    user_data = {}
+
+    user_fullname = request.form.get('fullname')
+    user_phone1 = request.form.get('phone1')
+    user_phone2 = request.form.get('phone2')
+    user_about = request.form.get('about')
+    user_address = request.form.get('address')
+    user_town = request.form.get('town')
+    user_city = request.form.get('city')
+    user_state = request.form.get('state')
+    user_photo = request.files.get('photo')
+    
+    pics_fn = save_picture(user_photo)
+    
+    user_data['fullname'] = user_fullname
+    user_data['phone1'] = user_phone1
+    user_data['phone2'] = user_phone2
+    user_data['about'] = user_about
+    user_data['address'] = user_address
+    user_data['town'] = user_town
+    user_data['city'] = user_city
+    user_data['state'] = user_state
+    user_data['photo'] = pics_fn
     try:
         with db:
+            auth_user = request.current_user
             user = db.find_user_by_id(user_id)
             if user is None:
                 abort(404)
+            if auth_user is not None:
+                if auth_user.user_id != user.user_id:
+                    return jsonify({"error": "Unauthorized"}), 400
             user.about = user_data['about']
             user.address = user_data['address']
             user.city = user_data['city']
             user.town = user_data['town']
             user.state = user_data['state']
-            user.verified = user_data['verified']
-            user.gender = user_data['gender']
             user.fullname = user_data['fullname']
             user.phone1 = user_data['phone1']
-            user.email = user_data['email']
+            user.phone2 = user_data['phone2']
+            user.photo = user_data['photo']
             db.save()
             return jsonify(user.to_dict()), 200
     except Exception as e:
+        print(e)
         abort(404)
